set up state, changeHandler, formHandler, and installed/imported MUI
