+++ client/package.json
@@ -5,6 +5,9 @@
     "proxy": "http://localhost:3001",
     "dependencies": {
         "@apollo/client": "^3.6.9",
+        "@emotion/react": "^11.10.4",
+        "@emotion/styled": "^11.10.4",
+        "@mui/material": "^5.10.5",
         "@testing-library/jest-dom": "^5.16.5",
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^14.4.3",
